@@ -1,54 +1,70 @@
 {
     "name": "@loopkit/javascript",
-    "version": "1.0.6",
-    "description": "LoopKit JavaScript SDK for tracking events and user analytics",
-    "main": "dist/loopkit.js",
+    "version": "1.1.0",
+    "description": "JavaScript SDK for LoopKit analytics platform",
+    "type": "module",
+    "main": "dist/loopkit.cjs.js",
     "module": "dist/loopkit.esm.js",
+    "browser": "dist/loopkit.js",
+    "unpkg": "dist/loopkit.min.js",
     "types": "dist/index.d.ts",
     "files": [
-        "dist",
-        "README.md"
+        "dist/",
+        "README.md",
+        "LICENSE"
     ],
     "scripts": {
-        "build": "rollup -c",
-        "build:watch": "rollup -c -w",
+        "build": "npm run clean && npm run build:ts && npm run build:rollup",
+        "build:ts": "tsc",
+        "build:rollup": "rollup -c",
+        "deploy:cdn": "npm run build && ./deploy.sh",
+        "clean": "rm -rf dist/",
+        "dev": "rollup -c -w",
         "test": "jest",
         "test:watch": "jest --watch",
-        "lint": "eslint src/**/*.js",
-        "format": "prettier --write src/**/*.js",
-        "prepare": "npm run build",
-        "deploy:cdn": "./deploy.sh"
+        "test:coverage": "jest --coverage",
+        "lint": "eslint src/**/*.ts",
+        "lint:fix": "eslint src/**/*.ts --fix",
+        "format": "prettier --write src/**/*.ts",
+        "type-check": "tsc --noEmit"
     },
     "keywords": [
         "analytics",
         "tracking",
-        "events",
         "loopkit",
         "javascript",
+        "typescript",
         "sdk"
     ],
-    "author": "LoopKit",
+    "author": "LoopKit Team",
     "license": "MIT",
     "devDependencies": {
         "@babel/core": "^7.23.0",
         "@babel/preset-env": "^7.23.0",
-        "@rollup/plugin-babel": "^6.0.0",
-        "@rollup/plugin-commonjs": "^25.0.0",
-        "@rollup/plugin-node-resolve": "^15.2.0",
-        "@rollup/plugin-replace": "^6.0.2",
-        "@rollup/plugin-terser": "^0.4.0",
-        "eslint": "^8.50.0",
+        "@rollup/plugin-babel": "^6.0.4",
+        "@rollup/plugin-node-resolve": "^15.2.3",
+        "@rollup/plugin-replace": "^5.0.5",
+        "@rollup/plugin-terser": "^0.4.4",
+        "@rollup/plugin-typescript": "^11.1.5",
+        "@types/jest": "^29.5.8",
+        "@types/node": "^20.9.2",
+        "@typescript-eslint/eslint-plugin": "^6.12.0",
+        "@typescript-eslint/parser": "^6.12.0",
+        "eslint": "^8.54.0",
         "jest": "^29.7.0",
         "jest-environment-jsdom": "^29.7.0",
-        "prettier": "^3.0.0",
-        "rollup": "^4.0.0"
+        "prettier": "^3.0.3",
+        "rollup": "^4.3.0",
+        "ts-jest": "^29.1.1",
+        "tslib": "^2.6.2",
+        "typescript": "^5.3.2"
     },
     "repository": {
         "type": "git",
-        "url": "git+https://github.com/loopkitai/javascript-sdk.git"
+        "url": "https://github.com/loopkit/javascript-sdk.git"
     },
     "bugs": {
-        "url": "https://github.com/loopkitai/javascript-sdk/issues"
+        "url": "https://github.com/loopkit/javascript-sdk/issues"
     },
-    "homepage": "https://github.com/loopkitai/javascript-sdk#readme"
+    "homepage": "https://github.com/loopkit/javascript-sdk#readme"
 }
