@@ -1,50 +1,51 @@
-const resolve = require('@rollup/plugin-node-resolve');
-const commonjs = require('@rollup/plugin-commonjs');
-const babel = require('@rollup/plugin-babel');
-const terser = require('@rollup/plugin-terser');
-const replace = require('@rollup/plugin-replace');
-const { readFileSync } = require('fs');
+import resolve from '@rollup/plugin-node-resolve';
+import babel from '@rollup/plugin-babel';
+import terser from '@rollup/plugin-terser';
+import replace from '@rollup/plugin-replace';
+import typescript from '@rollup/plugin-typescript';
+import { readFileSync } from 'fs';
 
 const pkg = JSON.parse(readFileSync('./package.json', 'utf8'));
+const version = pkg.version;
+const name = 'LoopKit';
 
-const banner = `/*!
- * LoopKit JavaScript SDK v${pkg.version}
- * (c) ${new Date().getFullYear()} LoopKit
- * Released under the MIT License
+const banner = `/**
+ * LoopKit JavaScript SDK v${version}
+ * (c) ${new Date().getFullYear()} LoopKit Team
+ * Released under the MIT License.
  */`;
 
 const baseConfig = {
-  input: 'src/index.js',
+  input: 'src/index.ts',
+  external: [],
   plugins: [
     replace({
-      __VERSION__: JSON.stringify(pkg.version),
+      __VERSION__: JSON.stringify(version),
       preventAssignment: true,
     }),
-    resolve(),
-    commonjs(),
-    babel.default({
+    typescript({
+      tsconfig: './tsconfig.json',
+      declaration: false, // We generate declarations separately with tsc
+      declarationMap: false,
+    }),
+    resolve({
+      browser: true,
+      preferBuiltins: false,
+    }),
+    babel({
       babelHelpers: 'bundled',
       exclude: 'node_modules/**',
-      presets: [
-        [
-          '@babel/preset-env',
-          {
-            targets: {
-              browsers: ['> 1%', 'last 2 versions', 'IE >= 11'],
-            },
-          },
-        ],
-      ],
+      extensions: ['.js', '.ts'],
     }),
   ],
 };
 
-module.exports = [
+export default [
   // ES Module build
   {
     ...baseConfig,
     output: {
-      file: 'dist/loopkit.esm.js',
+      file: pkg.module,
       format: 'es',
       banner,
     },
@@ -54,44 +55,39 @@ module.exports = [
   {
     ...baseConfig,
     output: {
-      file: 'dist/loopkit.cjs.js',
+      file: pkg.main,
       format: 'cjs',
       banner,
-      exports: 'default',
+      exports: 'named',
     },
   },
 
-  // UMD build for browsers
+  // UMD build (for browsers)
   {
     ...baseConfig,
     output: {
-      file: 'dist/loopkit.js',
+      file: pkg.browser,
       format: 'umd',
-      name: 'LoopKit',
+      name,
       banner,
-      exports: 'default',
     },
   },
 
-  // Minified UMD build for CDN
+  // Minified UMD build
   {
     ...baseConfig,
     plugins: [
       ...baseConfig.plugins,
-      terser.default({
+      terser({
         output: {
-          comments(node, comment) {
-            return comment.value.includes('LoopKit JavaScript SDK');
-          },
+          comments: false,
         },
       }),
     ],
     output: {
-      file: 'dist/loopkit.min.js',
+      file: pkg.unpkg,
       format: 'umd',
-      name: 'LoopKit',
-      banner,
-      exports: 'default',
+      name,
     },
   },
 ];
