@@ -1,28 +1,25 @@
 /**
- * ID generation utility
+ * ID generation utilities
  */
 export class IdGenerator {
   /**
    * Generate a unique session ID
-   * @returns {string} Session ID
    */
-  generateSessionId() {
+  generateSessionId(): string {
     return 'sess_' + Math.random().toString(36).substr(2, 9) + '_' + Date.now();
   }
 
   /**
    * Generate a unique anonymous ID
-   * @returns {string} Anonymous ID
    */
-  generateAnonymousId() {
+  generateAnonymousId(): string {
     return 'anon_' + Math.random().toString(36).substr(2, 9) + '_' + Date.now();
   }
 
   /**
    * Generate a unique event ID
-   * @returns {string} Event ID
    */
-  generateEventId() {
+  generateEventId(): string {
     return 'evt_' + Math.random().toString(36).substr(2, 9) + '_' + Date.now();
   }
 }
